pyproject.toml
[tool.poetry]
name = "djangoprojectcelerytest"
version = "0.1.0"
description = ""
authors = ["ig-rudenko <ig.rudenko1@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.8"
celery = "^5.4.0"
redis = "^5.0.8"
gunicorn = "^22.0.0"
django-celery-beat = "^2.6.0"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = {extras = ["crypto"], version = "^5.3.1"}
django-filter = "^24.3"


[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^4.4.6"
black = {extras = ["d"], version = "^24.8.0"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
